# v0.0.2/swagger.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Server"
    }
  ],
  "info": {
    "version": "0.0.1",
    "title": "v0.0.2",
    "description": "반창꼬 Server API 문서입니다."
  },
  "paths": {
    "/auth": {
      "post": {
        "summary": "로그인",
        "tags": ["Auth"],
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "auth.login",
        "x-nestia-jsDocTags": []
      }
    },
    "/mood": {
      "post": {
        "summary": "기분 등록",
        "tags": ["Mood"],
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoodDto"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "mood.insertMood",
        "x-nestia-jsDocTags": []
      },
      "delete": {
        "summary": "기분 삭제",
        "tags": ["Mood"],
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/__type"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "mood.deleteMood",
        "x-nestia-jsDocTags": []
      }
    },
    "/mood/daily/all/{date}": {
      "get": {
        "summary": "전 회원 일간 기분 평균 조회",
        "tags": ["Daily"],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Daily"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "mood.daily.all.getAllDailyMood",
        "x-nestia-jsDocTags": []
      }
    },
    "/mood/weekly/all/{date}": {
      "get": {
        "summary": "전 회원 주간 기분 평균 조회",
        "tags": ["Weekly"],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Weekly"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "mood.weekly.all.getAllWeeklyMood",
        "x-nestia-jsDocTags": []
      }
    },
    "/mood/daily/{date}": {
      "get": {
        "summary": "회원 일간 기분 조회",
        "tags": ["Daily"],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Daily"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "mood.daily.getDailyMood",
        "x-nestia-jsDocTags": []
      }
    },
    "/mood/weekly/{date}": {
      "get": {
        "summary": "회원 주간 기분 조회",
        "tags": ["Weekly"],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Weekly"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "mood.weekly.getWeeklyMood",
        "x-nestia-jsDocTags": []
      }
    },
    "/mood/relation": {
      "get": {
        "summary": "회원 기분 관계도 조회",
        "tags": ["Relation"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/WeatherMoodRelation"
                      },
                      {
                        "$ref": "#/components/schemas/WhatMoodRelation"
                      },
                      {
                        "$ref": "#/components/schemas/WhoMoodRelation"
                      }
                    ]
                  }
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "mood.relation.getRelations",
        "x-nestia-jsDocTags": []
      }
    },
    "/mood/{date}": {
      "get": {
        "summary": "기분 조회",
        "tags": ["Mood"],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "mood.getMood",
        "x-nestia-jsDocTags": []
      }
    },
    "/user": {
      "post": {
        "summary": "회원가입",
        "tags": ["User"],
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "user.createUser",
        "x-nestia-jsDocTags": []
      },
      "put": {
        "summary": "회원수정",
        "tags": ["User"],
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "user.updateUser",
        "x-nestia-jsDocTags": []
      },
      "delete": {
        "summary": "회원 삭제",
        "tags": ["User"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "user.deleteUser",
        "x-nestia-jsDocTags": []
      }
    },
    "/user/all": {
      "get": {
        "summary": "전 회원 닉네임 조회",
        "tags": ["User"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "user.all.getUserList",
        "x-nestia-jsDocTags": []
      }
    }
  },
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "description": "@format email",
            "x-typia-metaTags": [
              {
                "kind": "format",
                "value": "email"
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "format",
                "text": [
                  {
                    "text": "email",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "format": "email"
          },
          "password": {
            "description": "@pattern ^(?=.*[a-z])(?=.*[0-9])[a-zA-Z0-9!@#$%^&*_-]+$\n@minLength 8\n@maxLength 16",
            "x-typia-metaTags": [
              {
                "kind": "pattern",
                "value": "^(?=.*[a-z])(?=.*[0-9])[a-zA-Z0-9!@#$%^&*_-]+$"
              },
              {
                "kind": "minLength",
                "value": 8
              },
              {
                "kind": "maxLength",
                "value": 16
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "pattern",
                "text": [
                  {
                    "text": "^(?=.*[a-z])(?=.*[0-9])[a-zA-Z0-9!@#$%^&*_-]+$",
                    "kind": "text"
                  }
                ]
              },
              {
                "name": "minLength",
                "text": [
                  {
                    "text": "8",
                    "kind": "text"
                  }
                ]
              },
              {
                "name": "maxLength",
                "text": [
                  {
                    "text": "16",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "pattern": "^(?=.*[a-z])(?=.*[0-9])[a-zA-Z0-9!@#$%^&*_-]+$",
            "minLength": 8,
            "maxLength": 16
          },
          "autologin": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "boolean"
          }
        },
        "nullable": false,
        "required": [
          "email",
          "password",
          "autologin"
        ],
        "description": "@pattern ^(?=.*[a-z])(?=.*[0-9])[a-zA-Z0-9!@#$%^&*_-]+$\n@minLength 8\n@maxLength 16",
        "x-typia-jsDocTags": []
      },
      "MoodDto": {
        "type": "object",
        "properties": {
          "mood": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "weather": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "array",
            "items": {
              "x-typia-required": true,
              "x-typia-optional": false,
              "type": "number"
            }
          },
          "who": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "array",
            "items": {
              "x-typia-required": true,
              "x-typia-optional": false,
              "type": "number"
            }
          },
          "what": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "array",
            "items": {
              "x-typia-required": true,
              "x-typia-optional": false,
              "type": "number"
            }
          },
          "date": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string"
          }
        },
        "nullable": false,
        "required": [
          "mood",
          "weather",
          "who",
          "what",
          "date"
        ],
        "x-typia-jsDocTags": []
      },
      "Daily": {
        "type": "object",
        "properties": {
          "userId": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "year": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "weekNum": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "sun": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mon": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "tue": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "wed": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "thu": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "fri": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "sat": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "weekAvg": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          }
        },
        "nullable": false,
        "required": [
          "userId",
          "year",
          "weekNum",
          "sun",
          "mon",
          "tue",
          "wed",
          "thu",
          "fri",
          "sat",
          "weekAvg"
        ],
        "x-typia-jsDocTags": []
      },
      "Weekly": {
        "type": "object",
        "properties": {
          "userId": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "month": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string"
          },
          "week1": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "week2": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "week3": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "week4": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "week5": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "monthAvg": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          }
        },
        "nullable": false,
        "required": [
          "userId",
          "month",
          "week1",
          "week2",
          "week3",
          "week4",
          "week5",
          "monthAvg"
        ],
        "x-typia-jsDocTags": []
      },
      "WeatherMoodRelation": {
        "type": "object",
        "properties": {
          "userId": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "weatherId": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood1": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood2": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood3": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood4": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood5": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          }
        },
        "nullable": false,
        "required": [
          "userId",
          "weatherId",
          "mood1",
          "mood2",
          "mood3",
          "mood4",
          "mood5"
        ],
        "x-typia-jsDocTags": []
      },
      "WhatMoodRelation": {
        "type": "object",
        "properties": {
          "userId": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "whatId": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood1": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood2": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood3": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood4": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood5": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          }
        },
        "nullable": false,
        "required": [
          "userId",
          "whatId",
          "mood1",
          "mood2",
          "mood3",
          "mood4",
          "mood5"
        ],
        "x-typia-jsDocTags": []
      },
      "WhoMoodRelation": {
        "type": "object",
        "properties": {
          "userId": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "whoId": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood1": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood2": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood3": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood4": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          },
          "mood5": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "number"
          }
        },
        "nullable": false,
        "required": [
          "userId",
          "whoId",
          "mood1",
          "mood2",
          "mood3",
          "mood4",
          "mood5"
        ],
        "x-typia-jsDocTags": []
      },
      "__type": {
        "type": "object",
        "properties": {
          "date": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string"
          }
        },
        "nullable": false,
        "required": [
          "date"
        ],
        "x-typia-jsDocTags": []
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "email": {
            "description": "@format email",
            "x-typia-metaTags": [
              {
                "kind": "format",
                "value": "email"
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "format",
                "text": [
                  {
                    "text": "email",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "format": "email"
          },
          "nickName": {
            "description": "@pattern ^[a-zA-Z0-9가-힣_]+$\n@minLength 2\n@maxLength 8",
            "x-typia-metaTags": [
              {
                "kind": "pattern",
                "value": "^[a-zA-Z0-9가-힣_]+$"
              },
              {
                "kind": "minLength",
                "value": 2
              },
              {
                "kind": "maxLength",
                "value": 8
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "pattern",
                "text": [
                  {
                    "text": "^[a-zA-Z0-9가-힣_]+$",
                    "kind": "text"
                  }
                ]
              },
              {
                "name": "minLength",
                "text": [
                  {
                    "text": "2",
                    "kind": "text"
                  }
                ]
              },
              {
                "name": "maxLength",
                "text": [
                  {
                    "text": "8",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "pattern": "^[a-zA-Z0-9가-힣_]+$",
            "minLength": 2,
            "maxLength": 8
          },
          "password": {
            "description": "@pattern ^(?=.*[a-z])(?=.*[0-9])[a-zA-Z0-9!@#$%^&*_-]+$\n@minLength 8\n@maxLength 16",
            "x-typia-metaTags": [
              {
                "kind": "pattern",
                "value": "^(?=.*[a-z])(?=.*[0-9])[a-zA-Z0-9!@#$%^&*_-]+$"
              },
              {
                "kind": "minLength",
                "value": 8
              },
              {
                "kind": "maxLength",
                "value": 16
              }
            ],
            "x-typia-jsDocTags": [
              {
                "name": "pattern",
                "text": [
                  {
                    "text": "^(?=.*[a-z])(?=.*[0-9])[a-zA-Z0-9!@#$%^&*_-]+$",
                    "kind": "text"
                  }
                ]
              },
              {
                "name": "minLength",
                "text": [
                  {
                    "text": "8",
                    "kind": "text"
                  }
                ]
              },
              {
                "name": "maxLength",
                "text": [
                  {
                    "text": "16",
                    "kind": "text"
                  }
                ]
              }
            ],
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string",
            "pattern": "^(?=.*[a-z])(?=.*[0-9])[a-zA-Z0-9!@#$%^&*_-]+$",
            "minLength": 8,
            "maxLength": 16
          }
        },
        "nullable": false,
        "required": [
          "email",
          "nickName",
          "password"
        ],
        "description": "@pattern ^(?=.*[a-z])(?=.*[0-9])[a-zA-Z0-9!@#$%^&*_-]+$\n@minLength 8\n@maxLength 16",
        "x-typia-jsDocTags": []
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "apiKey",
        "name": "Autorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}